Non-working attempt to add a dropdown to the run configuration for Reveal
